// src/app/shared/interfaces/auxilium/inventory-center/product.interface.ts
export interface InventoryProductItem {
    id: number;
    itemcode: string;
    description: string;
    branchid: number;
    unitofmeasure: string;
    defaultquantity: number;
    taxable: boolean;
    isasset: boolean;
    vendor: string;
    vendorsku: string;
    itemdiagnosis: string;
    cmngroup1: string;
    cmngroup2: string;
    cmngroup3: string;
    cmngroup4: string;
    billingmultiplier: number;
    lastcost: number;
    averagecost: number;
    glr: string;
    glp: string;
    gladj: string;
    glcash: string;
    glasset: string;
    glcogs: string;
    inventorycategory: string;
    span: number;
    ndc: string;
    upc: string;
    note: string;
    adduserid: string;
    adddate: string;
    itemcategory: string;
    inactive: boolean;
    itemstatus: string;
    purchasemultiplier: number;
    purchaseunitofmeasure: string;
    hcpcscode: string;
    cbitem: boolean;
    description2: string;
    lot: boolean;
    upnqualifier: string;
    upn: string;
    level: string;
    discreetpackaging: boolean;
    audio: string;
    isfacilitybillitem: boolean;
    printdrworkorders: boolean;
    isepoitem: boolean;
    manufacturer: string;
    manufacturerpartno: string;
    make: string;
    model: string;
    alternatecode: string;
}

export type GetInventoryProductResponse = InventoryProductItem[];